refactor: move middlewares to middlewares folder
