+++ src/main/middlewares/index.ts
@@ -1,2 +1,2 @@
-export * from './webhook-validator/webhook-validator-middleware'
-export * from './auth/auth-nest-middleware'
+export * from './webhook-validator-middleware'
+export * from './auth-nest-middleware'
